refactor the SparseAttnIndexer as CustomOp
