--- vllm/attention/ops/rocm_aiter_mla_sparse.py
@@ -1,399 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-import importlib
-from functools import lru_cache
-
-import torch
-
-from vllm._aiter_ops import rocm_aiter_ops
-from vllm.logger import init_logger
-from vllm.platforms import current_platform
-from vllm.triton_utils import tl, triton
-
-logger = init_logger(__name__)
-
-
-@triton.jit
-def _indexer_k_quant_and_cache_kernel(
-    k_ptr,  # [num_tokens, head_dim]
-    kv_cache_ptr,  # [n_blks, blk_size//tile_block, head_dim // 16B, tile_block, 16B]
-    kv_cache_scale_ptr,  # [n_blks, blk_size]
-    slot_mapping_ptr,  # [num_tokens]
-    kv_cache_scale_stride,
-    kv_cache_value_stride,
-    block_size,
-    num_tokens,
-    head_dim: tl.constexpr,
-    BLOCK_TILE_SIZE: tl.constexpr,
-    HEAD_TILE_SIZE: tl.constexpr,
-    IS_FNUZ: tl.constexpr,
-    USE_UE8M0: tl.constexpr,
-):
-    tid = tl.program_id(0)
-    offset = tl.arange(0, head_dim)
-    tile_offset = (
-        offset // HEAD_TILE_SIZE * BLOCK_TILE_SIZE * HEAD_TILE_SIZE
-        + offset % HEAD_TILE_SIZE
-    )
-    tile_store_offset = tile_offset
-    # for idx in tl.range(tid, num_tokens, n_program):
-    src_ptr = k_ptr + tid * head_dim
-    slot_id = tl.load(slot_mapping_ptr + tid)
-    if slot_id < 0:
-        return
-    block_id = slot_id // block_size
-    block_offset = slot_id % block_size
-    tile_block_id = block_offset // BLOCK_TILE_SIZE
-    tile_block_offset = block_offset % BLOCK_TILE_SIZE
-    val = tl.load(src_ptr + offset)
-    amax = tl.max(val.abs(), axis=-1).to(tl.float32)
-    if IS_FNUZ:
-        scale = tl.maximum(1e-4, amax) / 224.0
-    else:
-        scale = tl.maximum(1e-4, amax) / 448.0
-
-    if USE_UE8M0:
-        scale = tl.exp2(tl.ceil(tl.log2(scale)))
-
-    fp8_val = (val.to(tl.float32) / scale).to(kv_cache_ptr.type.element_ty)
-    dst_ptr = (
-        kv_cache_ptr
-        + block_id * kv_cache_value_stride
-        + tile_block_id * BLOCK_TILE_SIZE * head_dim
-        + tile_block_offset * HEAD_TILE_SIZE
-    )
-    tl.store(dst_ptr + tile_store_offset, fp8_val)
-    dst_scale_ptr = kv_cache_scale_ptr + block_id * kv_cache_scale_stride + block_offset
-    tl.store(dst_scale_ptr, scale)
-
-
-def indexer_k_quant_and_cache_triton(
-    k: torch.Tensor,
-    kv_cache: torch.Tensor,  # [num_blocks, block_size, head_dim + 4]
-    slot_mapping: torch.Tensor,
-    quant_block_size,
-    scale_fmt,
-    block_tile_size=16,
-    head_tile_size=16,
-):
-    num_blocks = kv_cache.shape[0]
-    head_dim = k.shape[-1]
-    num_tokens = slot_mapping.shape[0]
-    block_size = kv_cache.shape[1]
-    # In real layout, we store the first portion as kv cache value
-    # and second portion as kv cache scale
-    kv_cache = kv_cache.view(num_blocks, -1)
-    kv_cache_value = kv_cache[:, : block_size * head_dim]
-    kv_cache_scale = kv_cache[:, block_size * head_dim :].view(torch.float32)
-    head_tile_size = head_tile_size // kv_cache.element_size()
-    grid = (num_tokens,)
-    _indexer_k_quant_and_cache_kernel[grid](
-        k,
-        kv_cache_value,
-        kv_cache_scale,
-        slot_mapping,
-        kv_cache_scale.stride(0),
-        kv_cache_value.stride(0),
-        block_size,
-        num_tokens,
-        head_dim,
-        block_tile_size,
-        head_tile_size,
-        IS_FNUZ=current_platform.fp8_dtype() == torch.float8_e4m3fnuz,
-        USE_UE8M0=scale_fmt == "ue8m0",
-    )
-
-
-@triton.jit
-def _cp_gather_indexer_quant_cache_kernel(
-    kv_cache_ptr,  # [n_blks,blk_size//tile_blk,head_dim//16B,tile_blk,16B]
-    kv_cache_scale_ptr,  # [n_blks, blk_size]
-    k_fp8_ptr,  # [num_tokens, head_dim]
-    k_scale_ptr,  # [num_tokens]
-    block_table_ptr,  # [batch_size, block_table_stride]
-    cu_seqlen_ptr,  # [batch_size + 1]
-    token_to_seq_ptr,  # [num_tokens]
-    block_size,
-    block_table_stride,
-    kv_cache_stride,
-    kv_cache_scale_stride,
-    HEAD_DIM: tl.constexpr,
-    BLOCK_TILE_SIZE: tl.constexpr,
-    HEAD_TILE_SIZE: tl.constexpr,
-):
-    tid = tl.program_id(0)
-    offset = tl.arange(0, HEAD_DIM)
-    batch_id = tl.load(token_to_seq_ptr + tid)
-    batch_start = tl.load(cu_seqlen_ptr + batch_id)
-    batch_end = tl.load(cu_seqlen_ptr + batch_id + 1)
-    batch_offset = tid - batch_start
-    if tid >= batch_end:
-        return
-    block_table_id = batch_offset // block_size
-    block_offset = batch_offset % block_size
-    block_table_offset = batch_id * block_table_stride + block_table_id
-    block_id = tl.load(block_table_ptr + block_table_offset)
-    tiled_block_id = block_offset // BLOCK_TILE_SIZE
-    tiled_block_offset = block_offset % BLOCK_TILE_SIZE
-    src_cache_offset = (
-        block_id * kv_cache_stride
-        + tiled_block_id * HEAD_DIM * BLOCK_TILE_SIZE
-        + tiled_block_offset * HEAD_TILE_SIZE
-    )
-    src_scale_offset = block_id * kv_cache_scale_stride + block_offset
-    dst_offset = tid * HEAD_DIM
-    src_scale_ptr = kv_cache_scale_ptr + src_scale_offset
-    src_cache_ptr = kv_cache_ptr + src_cache_offset
-    dst_k_ptr = k_fp8_ptr + dst_offset
-    scale_val = tl.load(src_scale_ptr)
-    tl.store(k_scale_ptr + tid, scale_val)
-    tiled_src_offset = (
-        offset // HEAD_TILE_SIZE * HEAD_TILE_SIZE * BLOCK_TILE_SIZE
-        + offset % HEAD_TILE_SIZE
-    )
-    val = tl.load(src_cache_ptr + tiled_src_offset)
-    tl.store(dst_k_ptr + offset, val)
-
-
-def cp_gather_indexer_k_quant_cache_triton(
-    k_cache: torch.Tensor,  # [num_blocks, block_size, head_dim + 4]
-    k_fp8: torch.Tensor,
-    k_fp8_scale: torch.Tensor,
-    block_table: torch.Tensor,
-    cu_seqlen: torch.Tensor,
-    token_to_seq: torch.Tensor,
-    block_tile_size: int = 16,
-    head_tile_size: int = 16,
-):
-    num_tokens = k_fp8.size(0)
-    block_size = k_cache.size(1)
-    block_table_stride = block_table.stride(0)
-    head_dim = k_fp8.shape[-1]
-    num_blocks = k_cache.shape[0]
-    # we assume the kv cache already been split to 2 portion
-    k_cache = k_cache.view(num_blocks, -1)
-    fp8_dtype = current_platform.fp8_dtype()
-    k_cache_value = k_cache[:, : block_size * head_dim].view(fp8_dtype)
-    k_cache_scale = k_cache[:, block_size * head_dim :].view(torch.float32)
-    grid = (num_tokens,)
-    k_fp8_scale = k_fp8_scale.view(torch.float32)
-    _cp_gather_indexer_quant_cache_kernel[grid](
-        k_cache_value,
-        k_cache_scale,
-        k_fp8,
-        k_fp8_scale,
-        block_table,
-        cu_seqlen,
-        token_to_seq,
-        block_size,
-        block_table_stride,
-        k_cache_value.stride(0),
-        k_cache_scale.stride(0),
-        head_dim,
-        block_tile_size,
-        head_tile_size,
-    )
-
-
-# Take from https://github.com/deepseek-ai/DeepGEMM/blob/main/tests/test_attention.py#L84
-def fp8_mqa_logits_torch(
-    q: torch.Tensor,
-    kv: tuple[torch.Tensor, torch.Tensor],
-    weights: torch.Tensor,
-    cu_seqlen_ks: torch.Tensor,
-    cu_seqlen_ke: torch.Tensor,
-) -> torch.Tensor:
-    """Compute FP8 MQA logits for a single sequence without KV paging.
-
-    Args:
-        q: Query tensor of shape [M, H, D]. Casted to
-            `torch.float8_e4m3fn` by caller.
-        kv: Tuple `(k_fp8, k_scales)` where `k_fp8` has shape [N, D] with
-            dtype `torch.float8_e4m3fn` and `k_scales` has shape [N] (or
-            [N, 1]) with dtype `torch.float32`.
-        weights: weights of shape [M, H], dtype `torch.float32`.
-        cu_seqlen_ks: Start indices (inclusive) for valid K per query position,
-            shape [M], dtype int32.
-        cu_seqlen_ke: End indices (exclusive) for valid K per query position,
-            shape [M], dtype int32.
-
-    Returns:
-        Logits tensor of shape [M, N], dtype `torch.float32`.
-    """
-    kv, scale = kv
-    seq_len_kv = kv.shape[0]
-    k = kv.to(torch.bfloat16)
-    q = q.to(torch.bfloat16)
-
-    mask_lo = (
-        torch.arange(0, seq_len_kv, device="cuda")[None, :] >= cu_seqlen_ks[:, None]
-    )
-    mask_hi = (
-        torch.arange(0, seq_len_kv, device="cuda")[None, :] < cu_seqlen_ke[:, None]
-    )
-    mask = mask_lo & mask_hi
-
-    score = torch.einsum("mhd,nd->hmn", q, k).float() * scale
-    logits = (score.relu() * weights.unsqueeze(-1).transpose(0, 1)).sum(dim=0)
-    logits = logits.masked_fill(~mask, float("-inf"))
-
-    return logits
-
-
-def rocm_fp8_mqa_logits(
-    q: torch.Tensor,
-    kv: tuple[torch.Tensor, torch.Tensor],
-    weights: torch.Tensor,
-    cu_seqlen_ks: torch.Tensor,
-    cu_seqlen_ke: torch.Tensor,
-) -> torch.Tensor:
-    """Compute FP8 MQA logits for a single sequence without KV paging.
-
-    Args:
-        q: Query tensor of shape [M, H, D]. Casted to
-            `torch.float8_e4m3fn` by caller.
-        kv: Tuple `(k_fp8, k_scales)` where `k_fp8` has shape [N, D] with
-            dtype `torch.float8_e4m3fn` and `k_scales` has shape [N] (or
-            [N, 1]) with dtype `torch.float32`.
-        weights: weights of shape [M, H], dtype `torch.float32`.
-        cu_seqlen_ks: Start indices (inclusive) for valid K per query position,
-            shape [M], dtype int32.
-        cu_seqlen_ke: End indices (exclusive) for valid K per query position,
-            shape [M], dtype int32.
-
-    Returns:
-        Logits tensor of shape [M, N], dtype `torch.float32`.
-    """
-
-    # TODO(ganyi): Temporarily workaround, will remove the module check and reference
-    # path after aiter merge this kernel into main
-    @lru_cache
-    def has_mqa_logits_module():
-        return importlib.util.find_spec("aiter.ops.triton.fp8_mqa_logits") is not None
-
-    if rocm_aiter_ops.is_enabled() and has_mqa_logits_module():
-        from aiter.ops.triton.fp8_mqa_logits import fp8_mqa_logits
-
-        kv, scale = kv
-        return fp8_mqa_logits(q, kv, scale, weights, cu_seqlen_ks, cu_seqlen_ke)
-    else:
-        return fp8_mqa_logits_torch(q, kv, weights, cu_seqlen_ks, cu_seqlen_ke)
-
-
-# Taken from https://github.com/deepseek-ai/DeepGEMM/blob/main/tests/test_attention.py#L156
-def fp8_paged_mqa_logits_torch(
-    q: torch.Tensor,
-    kv_cache: torch.Tensor,
-    weights: torch.Tensor,
-    context_lens: torch.Tensor,
-    block_tables: torch.Tensor,
-    max_model_len: int,
-):
-    from vllm.utils.math_utils import cdiv
-
-    fp8_dtype = current_platform.fp8_dtype()
-    batch_size, next_n, _, dim = q.size()
-    kv_cache, scale = kv_cache[..., :dim], kv_cache[..., dim:]
-    scale = scale.contiguous().view(torch.float)
-    q = q.float()
-    kv_cache = kv_cache.view(fp8_dtype).float() * scale
-    num_block, block_size, _, dim = kv_cache.size()
-    logits = torch.full(
-        [batch_size * next_n, max_model_len],
-        float("-inf"),
-        device=q.device,
-        dtype=torch.float32,
-    )
-    context_lens = context_lens.tolist()
-    for i in range(batch_size):
-        context_len = context_lens[i]
-        q_offsets = torch.arange(context_len - next_n, context_len, device="cuda")
-        weight_slice = (
-            weights[i * next_n : (i + 1) * next_n, :].transpose(0, 1).contiguous()
-        )
-        for block_rk in range(cdiv(context_len, block_size)):
-            block_idx = block_tables[i][block_rk]
-            qx, kx = q[i], kv_cache[block_idx]
-            k_offsets = torch.arange(
-                block_rk * block_size, (block_rk + 1) * block_size, device="cuda"
-            )
-            mask = (k_offsets[None, :] < context_len) & (
-                k_offsets[None, :] <= q_offsets[:, None]
-            )
-            s = torch.where(
-                mask[None, :, :],
-                (qx.transpose(0, 1) @ kx.transpose(0, 1).transpose(1, 2)).to(
-                    logits.dtype
-                ),
-                float("-inf"),
-            )
-            s = torch.relu(s) * weight_slice[..., None]
-            s = s.sum(dim=0)
-            logits[
-                i * next_n : (i + 1) * next_n,
-                block_rk * block_size : (block_rk + 1) * block_size,
-            ] = torch.where(k_offsets[None, :] <= q_offsets[:, None], s, float("-inf"))
-    return logits
-
-
-def rocm_fp8_paged_mqa_logits(
-    q_fp8: torch.Tensor,
-    kv_cache_fp8: torch.Tensor,
-    weights: torch.Tensor,
-    context_lens: torch.Tensor,
-    block_tables: torch.Tensor,
-    schedule_metadata: torch.Tensor,
-    max_model_len: int,
-) -> torch.Tensor:
-    """Compute FP8 MQA logits using paged KV-cache.
-
-    Args:
-        q_fp8: Query tensor of shape [B, next_n, H, D]. Casted to
-            `torch.float8_e4m3fn` by caller.
-        kv_cache_fp8: Paged KV-cache in packed FP8+scale layout with shape
-            [num_blocks, block_size, 1, D+4], dtype `torch.uint8`. The last
-            4 bytes per (block,pos) store the `float` dequant scale.
-        weights: Tensor of shape [B * next_n, H], dtype `torch.float32`.
-        context_lens: Tensor of shape [B], dtype int32; effective context length
-            for each batch element.
-        block_tables: Tensor of shape [B, max_blocks], dtype int32; maps logical
-            block indices to physical blocks in the paged cache.
-        schedule_metadata: Returned by `get_paged_mqa_logits_metadata`;
-            used to distribute work across SMs.
-        max_model_len: Maximum sequence length used to size the logits output.
-
-    Returns:
-        Logits tensor of shape [B * next_n, max_model_len], dtype
-        `torch.float32`.
-    """
-
-    if rocm_aiter_ops.is_enabled():
-        batch_size, next_n, heads, head_dim = q_fp8.shape
-        num_blocks, block_size, _, _ = kv_cache_fp8.shape
-
-        from aiter.ops.triton.pa_mqa_logits import deepgemm_fp8_paged_mqa_logits
-
-        out_logits = torch.full(
-            [batch_size * next_n, max_model_len],
-            float("-inf"),
-            device="cuda",
-            dtype=torch.float32,
-        )
-        deepgemm_fp8_paged_mqa_logits(
-            q_fp8,
-            kv_cache_fp8,
-            weights,
-            out_logits,
-            context_lens,
-            block_tables,
-            max_model_len,
-            ChunkK=256,
-            Preshuffle=block_size == 64,
-            KVBlockSize=block_size,
-            WavePerEU=2,
-        )
-        return out_logits
-    else:
-        return fp8_paged_mqa_logits_torch(
-            q_fp8, kv_cache_fp8, weights, context_lens, block_tables, max_model_len
-        )
